Player Farming Mechanic - WIP
The player now interacts with triggers setup in the Player Interaction Trigger (Layer 8) to find the closes one to a point in front of her.
This interaction become the target interaction if the player is to click (not yet implemented)
You can see a sphere jumping around in the scene (not the game) to visually see to adjust the range and how it affects the user experiance.