Weed Vines 1.0
First iteration of weed vines. They are a simple "physics joint rope" which isn't good enough for most games, but because our vines aren't moving much, nor is there going to be much physics-based interaction between these vines and the player, I'm sure this will prove sufficient.

We're going to need some art to be able to get the vines to their full beauty. I'm thinking a end "cap" and some leaves single and group variety to be assembled in Unity itself.

If the every art asset for the vines is rigged and exported with bones (no animations) that will allow for better IK and physics based animations. For example we could use some noise or Unity wind to generate dynamic animations for the leaves.