WIP - working on the watering can
Free Watering Can Model added
Watering Can Shader added
Need a splash particle effect for when it's full
Need to have the player be able to cycle to it
Want the player to carry it ik to their hand: probably just toggle visibility of the watering can model?
In the Plant's SeedState, I'm setting it up to just trigger the next state.
Removed extra files added during player controller fix